1.1.1 Bug fix release.
git-svn-id: https://plugins.svn.wordpress.org/backupwordpress/trunk@367199 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ functions/backup.mysql.functions.php
@@ -72,8 +72,8 @@ function hmbkp_mysqldump_path() {
 
 	endif;
 
-	// Check again incase the saved path has stopped working for some reason
-	if ( !shell_exec( 'which ' . $path ) ) :
+	// Check again in-case the saved path has stopped working for some reason
+	if ( $path && !shell_exec( 'which ' . $path ) ) :
 		delete_option( 'hmbkp_mysqldump_path' );
 		return hmbkp_mysqldump_path();
